sandbox: replace transparent-canvas experiments with scene-background (CPU + GPU)
The transparent-canvas experiments demonstrated the (unsolvable-in-shader)
problem; remove them. Keep the working solution — additive over a gradient
scene.background on an opaque canvas — as CPU (SpriteRenderer) and GPU
(GPURenderer) experiments sharing one particle system.
