+++ sandbox/experiments/additive-blending-scene-background-gpu/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Additive Blending — Scene Background (GPU)</title>
+    <meta charset="UTF-8" />
+    <link rel="stylesheet" type="text/css" href="/style/reset.css" />
+    <link rel="stylesheet" type="text/css" href="/style/app.css" />
+  </head>
+
+  <body>
+    <div id="app">
+      <canvas id="canvas"></canvas>
+      <div id="description">
+        GPURenderer, additive no-alpha texture (<code>circle_01.png</code>) with
+        the gradient rendered <strong>in the scene</strong>
+        (<code>scene.background</code>) on an opaque canvas. Same system as the
+        CPU experiment — clean library, no shader hacks, additive Just Works.
+      </div>
+    </div>
+
+    <script type="module" src="./index.js"></script>
+  </body>
+</html>
