--- sandbox/experiments/transparent-canvas-additive-blending/index.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Transparent Canvas Additive Blending</title>
-    <meta charset="UTF-8" />
-    <link rel="stylesheet" type="text/css" href="/style/reset.css" />
-    <link rel="stylesheet" type="text/css" href="/style/app.css" />
-    <style>
-      /* The WebGL canvas is cleared transparent (renderer clearAlpha = 0), so
-         this CSS gradient painted on the canvas element shows through — the
-         same situation as an `alpha: true` host whose page background is a CSS
-         gradient. See specs/fix-133-additive-transparent-canvas.md. */
-      #canvas {
-        background: linear-gradient(
-          160deg,
-          #f6d365 0%,
-          #fda085 45%,
-          #a18cd1 100%
-        );
-      }
-    </style>
-  </head>
-
-  <body>
-    <div id="app">
-      <canvas id="canvas"></canvas>
-      <div id="description">
-        #133 repro — an additive, no-alpha texture (<code>circle_01.png</code>,
-        opaque black background) on an <code>alpha: true</code> canvas over a CSS
-        gradient. Particles render as opaque black squares instead of
-        compositing additively.
-      </div>
-    </div>
-
-    <script type="module" src="./index.js"></script>
-  </body>
-</html>
