+++ sandbox/index.html
@@ -11,13 +11,13 @@
     <nav>
       <ul>
         <li>
-          <a href="experiments/transparent-canvas-additive-blending/index.html">
-            Transparent Canvas Additive Blending
+          <a href="experiments/additive-blending-scene-background-cpu/index.html">
+            Additive Blending — Scene Background (CPU)
           </a>
         </li>
         <li>
-          <a href="experiments/transparent-canvas-additive-blending-gpu/index.html">
-            Transparent Canvas Additive Blending (GPU)
+          <a href="experiments/additive-blending-scene-background-gpu/index.html">
+            Additive Blending — Scene Background (GPU)
           </a>
         </li>
         <li>
